# docs/openapi/replicadb-server.json
{
  "components": {
    "responses": {
      "BadRequestProblem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description": "The request is malformed or fails validation."
      },
...
      "ForbiddenProblem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description": "The authenticated identity lacks the required permission."
      },
...
      "NotFoundProblem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description": "The requested resource does not exist or is not visible."
      },
...
      "UnauthorizedProblem": {
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemDetail"
            }
          }
        },
        "description": "Authentication is required or credentials are invalid."
      }
    },
    "schemas": {
...
      "JobScheduleRequest": {
        "description": "Recurring Quartz schedule for one job.",
        "properties": {
          "cronExpression": {
            "description": "Quartz CRON expression with seconds, minutes, hours, day, month, and weekday fields.",
            "example": "0 0 2 * * ?",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "description": "Whether the scheduler may create runs from this schedule.",
            "type": "boolean"
          },
          "timeZone": {
            "description": "IANA time-zone identifier. Null or blank defaults to UTC.",
            "example": "UTC",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "cronExpression"
        ],
        "type": "object"
      },
      "JobScheduleResponse": {
        "description": "Persisted job schedule and its currently computed next fire time.",
        "properties": {
          "createdAt": {
            "description": "Creation timestamp in UTC.",
            "format": "date-time",
            "type": "string"
          },
          "cronExpression": {
            "description": "Validated Quartz CRON expression.",
            "type": "string"
          },
          "enabled": {
            "description": "Whether the schedule is enabled.",
            "type": "boolean"
          },
          "jobDefinitionId": {
            "description": "Scheduled job definition identifier.",
            "format": "uuid",
            "type": "string"
          },
          "nextFireTime": {
            "description": "Next fire time computed by the current scheduler, or null when disabled or unscheduled.",
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "timeZone": {
            "description": "IANA time-zone identifier used by Quartz.",
            "type": "string"
          },
          "updatedAt": {
            "description": "Last update timestamp in UTC.",
            "format": "date-time",
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "RFC 7807 problem response. Dynamic detail text is credential-redacted.",
        "properties": {
          "detail": {
            "type": "string"
          },
          "instance": {
            "format": "uri-reference",
            "type": "string"
          },
          "status": {
            "format": "int32",
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "format": "uri-reference",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "csrfHeader": {
        "description": "CSRF header for protected state-changing requests. Its value matches the XSRF-TOKEN cookie.",
        "in": "header",
        "name": "X-XSRF-TOKEN",
        "type": "apiKey"
      },
      "sessionCookie": {
        "description": "Session cookie issued by an authenticated, same-origin ReplicaDB Server deployment.",
        "in": "cookie",
        "name": "JSESSIONID",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Static contract for the authenticated ReplicaDB Server control plane. Use it against a same-origin deployment that owns the session and CSRF cookies.",
    "title": "ReplicaDB Server API",
    "version": "v1"
  },
  "openapi": "3.0.1",
  "paths": {
...
    "/api/v1/jobs/{jobDefinitionId}/schedule": {
      "delete": {
        "description": "Removes the persisted schedule and Quartz trigger after EDIT permission. Repeated deletion is idempotent. Protected mutations require the session cookie and CSRF header.",
        "operationId": "deleteJobSchedule",
        "parameters": [
          {
            "description": "Job definition identifier.",
            "in": "path",
            "name": "jobDefinitionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Schedule removed or already absent"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundProblem"
          }
        },
        "security": [
          {
            "csrfHeader": [],
            "sessionCookie": []
          }
        ],
        "summary": "Delete a job schedule",
        "tags": [
          "Schedules"
        ]
      },
      "get": {
        "description": "Returns the recurring schedule and next fire time after VIEW permission. A job without a schedule returns not found.",
        "operationId": "getJobSchedule",
        "parameters": [
          {
            "description": "Job definition identifier.",
            "in": "path",
            "name": "jobDefinitionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/JobScheduleResponse"
                }
              }
            },
            "description": "Schedule returned"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundProblem"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "summary": "Get a job schedule",
        "tags": [
          "Schedules"
        ]
      },
      "put": {
        "description": "Validates and stores the Quartz CRON schedule after EDIT permission, then reconciles the durable scheduler trigger. Blank time zones default to UTC. Protected mutations require the session cookie and CSRF header.",
        "operationId": "upsertJobSchedule",
        "parameters": [
          {
            "description": "Job definition identifier.",
            "in": "path",
            "name": "jobDefinitionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/JobScheduleResponse"
                }
              }
            },
            "description": "Schedule created or replaced"
          },
          "400": {
            "$ref": "#/components/responses/BadRequestProblem"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedProblem"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenProblem"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundProblem"
          }
        },
        "security": [
          {
            "csrfHeader": [],
            "sessionCookie": []
          }
        ],
        "summary": "Create or replace a job schedule",
        "tags": [
          "Schedules"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Generated server url",
      "url": "http://localhost"
    }
  ],
...
}